implement a force_ssl option
This works similarly to Rails force_ssl. Why add this to a proxy gem when it's
stuff that generally lives in the webserver configuration? The most common
use case for using a proxy gem in the first place is not having full control
of the webserver (such as on heroku). In those environments the redirect is
generally configured to be done on Rails, but requests handled by rack reverse
are not handled by a Rails controller.
Ideally one would create another Rack middleware to do this, but I think the
convenience is worth it (especially in a Rails environment).